refactor: according added store
