+++ src/components/Modals/NewGameModal.tsx
@@ -1,12 +1,13 @@
 import React, { ChangeEventHandler, FormEventHandler, useState } from 'react';
 import Select from 'react-select';
 import { useHistory } from 'react-router-dom';
+import { useDispatch } from 'react-redux';
 import { dealerOptions, voteOptions } from '@/mocks/options';
 import { Input } from '@/components/Input';
 import { Modal } from '@/components/Modal';
 import styles from './Modals.module.scss';
-import { useAppDispatch } from '@/redux/reduxHooks';
-import { addGameSettings } from '@/redux/slice';
+import { AppDispatch } from '@/store';
+import { addGameSettings } from '@/store/actions/userActions';
 
 interface INewGameModalProps {
   show: boolean;
@@ -25,7 +26,7 @@ export const NewGameModal: React.FC<INewGameModalProps> = ({ show, onClick }) =>
     dealerRights: dealerOptions[0].value || undefined,
   });
 
-  const dispatch = useAppDispatch();
+  const dispatch = useDispatch<AppDispatch>();
   const history = useHistory();
 
   const inputsHandler: ChangeEventHandler<HTMLInputElement> = (event) => {
