+++ src/store/reducers/gameSettingsReducer.ts
@@ -11,8 +11,8 @@ export const gameSettings = createReducer(initialStateGame, (builder) => {
   builder
     .addCase(addGameSettings, (state, action) => {
       state.gameName = action.payload.gameName;
-      state.voteSystem = action.payload.voteSystem;
-      state.dealerRights = action.payload.dealerRights;
+      state.voteSystem = action.payload.voteSystem || '';
+      state.dealerRights = action.payload.dealerRights || '';
     })
     .addDefaultCase((state) => state);
 });
